test(brag): refactor integration-tests into sub-modules and add brag IT with LLM smoke tests
Split integration-tests into packaging=pom with sub-modules (brag, cognition,
google, terminal, smoke). Add brag IT sub-module with 30 unit tests and 2 LLM
integration tests. Remove @TestProfile and @Tag annotations. Add failsafe
plugin with explicit execution goals. Delete old flat test structure.

Closes #57
