--- integration-tests/api-llm-marker.txt
@@ -1,1 +1,0 @@
-api-file-test
