+++ integration-tests/brag/pom.xml
@@ -10,14 +10,19 @@
     <version>0.4.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>qlawkus-integration-tests-cognition</artifactId>
+  <artifactId>qlawkus-integration-tests-brag</artifactId>
   <packaging>quarkus</packaging>
 
   <dependencies>
     <dependency>
       <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-tools-brag</artifactId>
+      <version>0.3.0</version>
+    </dependency>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
       <artifactId>qlawkus-client</artifactId>
-      <version>${project.version}</version>
+      <version>0.4.0</version>
     </dependency>
     <dependency>
       <groupId>dev.omatheusmesmo</groupId>
@@ -36,6 +41,11 @@
       <scope>test</scope>
     </dependency>
     <dependency>
+      <groupId>io.rest-assured</groupId>
+      <artifactId>rest-assured</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
       <groupId>org.awaitility</groupId>
       <artifactId>awaitility</artifactId>
       <scope>test</scope>
@@ -49,9 +59,26 @@
         <artifactId>quarkus-maven-plugin</artifactId>
         <extensions>true</extensions>
       </plugin>
-      <plugin>
-        <artifactId>maven-surefire-plugin</artifactId>
-      </plugin>
+    <plugin>
+      <artifactId>maven-surefire-plugin</artifactId>
+    </plugin>
+    <plugin>
+      <artifactId>maven-failsafe-plugin</artifactId>
+      <configuration>
+        <systemPropertyVariables>
+          <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
+          <maven.home>${maven.home}</maven.home>
+        </systemPropertyVariables>
+      </configuration>
+      <executions>
+        <execution>
+          <goals>
+            <goal>integration-test</goal>
+            <goal>verify</goal>
+          </goals>
+        </execution>
+      </executions>
+    </plugin>
     </plugins>
   </build>
 </project>
