+++ integration-tests/brag/src/main/resources/application.properties
@@ -1,0 +1,21 @@
+quarkus.security.users.embedded.enabled=true
+quarkus.security.users.embedded.plain-text=true
+quarkus.security.users.embedded.users.qlawkus=qlawkus-test
+quarkus.security.users.embedded.roles.qlawkus=admin
+quarkus.http.auth.basic=true
+
+qlawkus.brag.enabled=true
+
+quarkus.datasource."brag".db-kind=postgresql
+quarkus.hibernate-orm."brag".datasource=brag
+quarkus.hibernate-orm."brag".physical-naming-strategy=org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy
+quarkus.flyway."brag".migrate-at-start=true
+quarkus.flyway."brag".baseline-on-migrate=true
+quarkus.flyway."brag".locations=db/migration-brag
+
+%test.qlawkus.startup-thought.enabled=false
+%test.quarkus.hibernate-orm.schema-management.strategy=validate
+%test.quarkus.flyway.clean-at-start=true
+%test.quarkus.flyway."brag".clean-at-start=true
+%test.quarkus.langchain4j.ollama."ollama-fallback".base-url=http://localhost:11434
+%test.quarkus.langchain4j.openai."nvidia".timeout=120s
