+++ integration-tests/google/pom.xml
@@ -1,76 +1,87 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
- xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
- xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
- <modelVersion>4.0.0</modelVersion>
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
 
- <parent>
-  <groupId>dev.omatheusmesmo</groupId>
-  <artifactId>qlawkus-integration-tests</artifactId>
-  <version>0.4.0</version>
- </parent>
+  <parent>
+    <groupId>dev.omatheusmesmo</groupId>
+    <artifactId>qlawkus-integration-tests</artifactId>
+    <version>0.4.0-SNAPSHOT</version>
+  </parent>
 
- <artifactId>qlawkus-integration-tests-google</artifactId>
- <packaging>quarkus</packaging>
+  <artifactId>qlawkus-integration-tests-google</artifactId>
+  <packaging>quarkus</packaging>
 
- <dependencies>
-  <dependency>
-   <groupId>dev.omatheusmesmo</groupId>
-   <artifactId>qlawkus-tools-google-auth</artifactId>
-   <version>${project.version}</version>
-  </dependency>
-  <dependency>
-   <groupId>io.quarkus</groupId>
-   <artifactId>quarkus-rest-client-jackson</artifactId>
-  </dependency>
-  <dependency>
-   <groupId>dev.omatheusmesmo</groupId>
-   <artifactId>qlawkus-tools-google-calendar</artifactId>
-   <version>${project.version}</version>
-  </dependency>
-  <dependency>
-   <groupId>dev.omatheusmesmo</groupId>
-   <artifactId>qlawkus-tools-google-gmail</artifactId>
-   <version>${project.version}</version>
-  </dependency>
-  <dependency>
-   <groupId>dev.omatheusmesmo</groupId>
-   <artifactId>qlawkus-tools-google-drive</artifactId>
-   <version>${project.version}</version>
-  </dependency>
-  <dependency>
-   <groupId>dev.omatheusmesmo</groupId>
-   <artifactId>qlawkus-tools-google-sheets</artifactId>
-   <version>${project.version}</version>
-  </dependency>
-  <dependency>
-   <groupId>dev.omatheusmesmo</groupId>
-   <artifactId>qlawkus-tools-google-storage</artifactId>
-   <version>${project.version}</version>
-  </dependency>
-  <dependency>
-   <groupId>io.quarkiverse.wiremock</groupId>
-   <artifactId>quarkus-wiremock</artifactId>
-   <version>${quarkus-wiremock.version}</version>
-   <scope>test</scope>
-  </dependency>
-  <dependency>
-   <groupId>io.quarkiverse.wiremock</groupId>
-   <artifactId>quarkus-wiremock-test</artifactId>
-   <version>${quarkus-wiremock.version}</version>
-   <scope>test</scope>
-  </dependency>
- </dependencies>
+  <dependencies>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-client</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-tools-google-auth</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.quarkus</groupId>
+      <artifactId>quarkus-rest-client-jackson</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-tools-google-calendar</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-tools-google-gmail</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-tools-google-drive</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-tools-google-sheets</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-tools-google-storage</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.quarkus</groupId>
+      <artifactId>quarkus-junit</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.quarkiverse.wiremock</groupId>
+      <artifactId>quarkus-wiremock</artifactId>
+      <version>${quarkus-wiremock.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.quarkiverse.wiremock</groupId>
+      <artifactId>quarkus-wiremock-test</artifactId>
+      <version>${quarkus-wiremock.version}</version>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
 
- <build>
-  <plugins>
-   <plugin>
-    <groupId>${quarkus.platform.group-id}</groupId>
-    <artifactId>quarkus-maven-plugin</artifactId>
-   </plugin>
-   <plugin>
-    <artifactId>maven-surefire-plugin</artifactId>
-   </plugin>
-  </plugins>
- </build>
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>${quarkus.platform.group-id}</groupId>
+        <artifactId>quarkus-maven-plugin</artifactId>
+        <extensions>true</extensions>
+      </plugin>
+      <plugin>
+        <artifactId>maven-surefire-plugin</artifactId>
+      </plugin>
+    </plugins>
+  </build>
 </project>
