+++ integration-tests/pom.xml
@@ -9,120 +9,13 @@
   </parent>
 
   <artifactId>qlawkus-integration-tests</artifactId>
-  <packaging>quarkus</packaging>
+  <packaging>pom</packaging>
 
-  <dependencies>
-        <dependency>
-            <groupId>dev.omatheusmesmo</groupId>
-            <artifactId>qlawkus-client</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>dev.omatheusmesmo</groupId>
-            <artifactId>qlawkus-tools-google-auth</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-rest-client-jackson</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>dev.omatheusmesmo</groupId>
-            <artifactId>qlawkus-tools-google-calendar</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>dev.omatheusmesmo</groupId>
-            <artifactId>qlawkus-tools-google-gmail</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>dev.omatheusmesmo</groupId>
-            <artifactId>qlawkus-tools-google-drive</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>dev.omatheusmesmo</groupId>
-            <artifactId>qlawkus-tools-google-sheets</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-      <dependency>
-        <groupId>dev.omatheusmesmo</groupId>
-        <artifactId>qlawkus-tools-google-storage</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>dev.omatheusmesmo</groupId>
-        <artifactId>qlawkus-tools-brag</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-junit</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-junit-mockito</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkiverse.wiremock</groupId>
-      <artifactId>quarkus-wiremock</artifactId>
-      <version>${quarkus-wiremock.version}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkiverse.wiremock</groupId>
-      <artifactId>quarkus-wiremock-test</artifactId>
-      <version>${quarkus-wiremock.version}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.rest-assured</groupId>
-      <artifactId>rest-assured</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.awaitility</groupId>
-      <artifactId>awaitility</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>${quarkus.platform.group-id}</groupId>
-        <artifactId>quarkus-maven-plugin</artifactId>
-        <extensions>true</extensions>
-      </plugin>
-    <plugin>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <configuration>
-          <systemPropertyVariables>
-            <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
-            <maven.home>${maven.home}</maven.home>
-          </systemPropertyVariables>
-        </configuration>
-      </plugin>
-    <plugin>
-      <artifactId>maven-failsafe-plugin</artifactId>
-      <configuration>
-        <systemPropertyVariables>
-          <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
-          <maven.home>${maven.home}</maven.home>
-        </systemPropertyVariables>
-      </configuration>
-      <executions>
-        <execution>
-          <goals>
-            <goal>integration-test</goal>
-            <goal>verify</goal>
-          </goals>
-        </execution>
-      </executions>
-    </plugin>
-    </plugins>
-  </build>
+  <modules>
+    <module>cognition</module>
+    <module>google</module>
+    <module>terminal</module>
+    <module>smoke</module>
+    <module>brag</module>
+  </modules>
 </project>
