+++ integration-tests/terminal/pom.xml
@@ -1,27 +1,47 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
- xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
- xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
- <modelVersion>4.0.0</modelVersion>
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
 
- <parent>
-  <groupId>dev.omatheusmesmo</groupId>
-  <artifactId>qlawkus-integration-tests</artifactId>
-  <version>0.4.0</version>
- </parent>
+  <parent>
+    <groupId>dev.omatheusmesmo</groupId>
+    <artifactId>qlawkus-integration-tests</artifactId>
+    <version>0.4.0-SNAPSHOT</version>
+  </parent>
 
- <artifactId>qlawkus-integration-tests-terminal</artifactId>
- <packaging>quarkus</packaging>
+  <artifactId>qlawkus-integration-tests-terminal</artifactId>
+  <packaging>quarkus</packaging>
 
- <build>
-  <plugins>
-   <plugin>
-    <groupId>${quarkus.platform.group-id}</groupId>
-    <artifactId>quarkus-maven-plugin</artifactId>
-   </plugin>
-   <plugin>
-    <artifactId>maven-surefire-plugin</artifactId>
-   </plugin>
-  </plugins>
- </build>
+  <dependencies>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-client</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>dev.omatheusmesmo</groupId>
+      <artifactId>qlawkus-testing-internal</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.quarkus</groupId>
+      <artifactId>quarkus-junit</artifactId>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
+
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>${quarkus.platform.group-id}</groupId>
+        <artifactId>quarkus-maven-plugin</artifactId>
+        <extensions>true</extensions>
+      </plugin>
+      <plugin>
+        <artifactId>maven-surefire-plugin</artifactId>
+      </plugin>
+    </plugins>
+  </build>
 </project>
